requirements.txt
click==8.1.7
pyyaml==6.0.2
pytest==8.3.3
kubernetes==31.0.0